move route from hoc screen-switch to app
